% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poLCA.bvr.R
\name{bvr}
\alias{bvr}
\title{Bivariate residuals for latent class models}
\usage{
bvr(fit, tol = 0.001, rescale_to_df = TRUE)
}
\arguments{
\item{fit}{A poLCA fit object}

\item{tol}{Optional: tolerance for small expected counts}

\item{rescale_to_df}{Optional: whether to divide the pairwise "chi-square" values by 
the degrees of freedom of the local crosstable. Default is TRUE.}
}
\value{
The table of bivariate residuals
}
\description{
Calculate the "bivariate residuals" (BVRs) between pairs of variables 
in a latent class model.
}
\details{
This function compares the model-implied (expected) counts in the crosstables
of all pairs of observed dependent variables to the observed counts. For each
pair, it calculates a "chi-square" statistic,

\deqn{\text{BVR} = \sum_{j, j'} \frac{(n_{jj'} - e_{jj'})^2}{e_{jj'}}},

where \eqn{n_{jj'}} are the observed counts for categories \eqn{j} and \eqn{j'} 
of the variables being crosstabulated, and \eqn{e_{jj'}} are
the expected counts under the latent class model.

Note that the BVR does not follow an asymptotic chi-square distribution and
for accurate p-values, parametric bootstrapping is necessary (Oberski et al. 2013).
}
\examples{
data(values)
f <- cbind(A, B, C, D) ~ 1
M0 <- poLCA(f,values, nclass=1, verbose = FALSE) 
bvr(M0) # 12.4, 5.7, 8.3, 15.6, ... 
}
\references{
Oberski, DL, Van Kollenburg, GH and Vermunt, JK (2013). 
  A Monte Carlo evaluation of three methods to detect local dependence in binary data latent class models. 
  Advances in Data Analysis and Classification 7 (3), 267-279.
}
\seealso{
\code{\link{poLCA}} for fitting the latent class model.
}
\author{
Daniel Oberski (daniel.oberski@gmail.com)
}
