\name{print.poLCA}
\alias{print.poLCA}
\title{Output an object of class poLCA}
\description{
Produces nice output of an object of class poLCA as generated by \code{\link{poLCA}}.
}
\usage{
print.poLCA(x, ...)
}
\arguments{
  \item{x}{An object of class poLCA.}
  \item{...}{}
}
\seealso{ \code{\link{poLCA}} }
\examples{
data(values)
f <- cbind(A,B,C,D)~1
M1 <- poLCA(f,values,nclass=2,verbose=FALSE)
print(M1)
}
\keyword{print}
