\name{poLCA-package}
\alias{poLCA-package}
\docType{package}
\title{Latent class analysis and latent class regression models for polytomous outcome variables}
\description{This package estimates latent class and latent class regression models for polytomous outcome variables.}
\details{
\tabular{ll}{
Package: \tab poLCA\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2007-04-04\cr
License: \tab GPL\cr
}
The function \code{\link{poLCA}} estimates the latent class model.
}
\author{
Drew Linzer \email{dlinzer@ucla.edu} and Jeffrey Lewis \email{jblewis@ucla.edu}

Maintainer: Drew Linzer \email{dlinzer@ucla.edu} \cr
Website: \url{http://dlinzer.bol.ucla.edu/poLCA}
}
\keyword{documentation}
