\name{poLCA-package}
\alias{poLCA-package}
\docType{package}
\title{Latent class analysis and latent class regression models for polytomous outcome variables}
\description{This package estimates latent class and latent class regression models for polytomous outcome variables.}
\details{
\tabular{ll}{
Package: \tab poLCA\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2007-11-01\cr
License: \tab GPL (>= 2)\cr
}
The function \code{\link{poLCA}} estimates the latent class model.
}
\author{
Drew Linzer \email{dlinzer@emory.edu} and Jeffrey Lewis \email{jblewis@ucla.edu}

Maintainer: Drew Linzer \email{dlinzer@emory.edu} \cr
Website: \url{http://userwww.service.emory.edu/~dlinzer/poLCA}
}
\keyword{documentation}
